+++ doc/source/changelog.md
@@ -55,6 +55,7 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Fixed graph not always saving when saving the scene
         - Fixed shader generator crash when a node has an unconnected input
         - Fixed cellular noise when used on GPU
+    - `VoxelInstancer`: fixed crash when hiding the node in the editor
     - `VoxelInstanceLibrary`: fixed `find_item_by_name` was not finding items
     - `VoxelLodTerrain`: Fixed error spam when re-generating or destroying the terrain
     - `VoxelStreamRegionFiles`: Fixed `block_size_po2` wasn't working correctly